% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{dataLevels}
\alias{dataLevels}
\title{Levels of a dataset}
\usage{
dataLevels(data.mat)
}
\arguments{
\item{data.mat}{A numeric matrix}
}
\value{
A data frame containing counts for each level of the hypercube
}
\description{
get the representation of different hypercube levels in a dataset
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
dataLevels(observations)
}
