% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HyperTraPS}
\alias{HyperTraPS}
\title{Runs HyperTraPS-related inference on a dataset of observations}
\usage{
HyperTraPS(
  obs,
  initialstates = NULL,
  priors = NULL,
  starttimes = NULL,
  endtimes = NULL,
  length = 3L,
  kernel = 5L,
  samplegap = -1L,
  losses = 0L,
  apm_type = 0L,
  sa = 0L,
  sgd = 0L,
  sgd_scale = 0.01,
  seed = 1L,
  outputinput = 0L,
  regularise = 0L,
  penalty = 0L,
  lasso = 0L,
  model = 2L,
  pli = 0L,
  walkers = 200L,
  full_analysis = 1L,
  limited_output = 0L,
  output_transitions = 1L,
  samples_per_row = 10L,
  featurenames = NULL
)
}
\arguments{
\item{initialstates}{An optional matrix of initial states. If we are using longitudinal observations, each row in this matrix gives the "before" state to the corresponding "after" state in the observations matrix. Omitting this matrix is equivalent to consider every observation to have a root "before" state. If specified, should be $n \times L$, containing $n$ cross-sectional observations of length $L$, to match the observations matrix.}

\item{losses}{Whether to consider accumulation of gains (0) or losses (1)}

\item{apm_type}{APM}

\item{sgd_scale}{SGD}

\item{seed}{Random number seed}

\item{outputinput}{Option to output the input data}

\item{regularise}{Regularise}

\item{model}{Model structure}

\item{pli}{Phenotype landscape inference}

\item{matrix}{A matrix of observations. Should contain 0s, 1s, and optional 2s for missing data. Should be $n \times L$, containing $n$ cross-sectional observations of length $L$.}

\item{start_times}{An optional vector of $n$ times describing the beginning of the observation time window for each observation. If empty, equivalent to this time window beginning at time 0. If specified, should be of length $n$.}

\item{end_times}{An optional vector of $n$ times describing the end of the observation time window for each observation. If empty, equivalent to this time window ending at time infinity. If specified, should be of length $n$.}

\item{length_index}{Length of MCMC chain}

\item{kernel_index}{Kernel index}
}
\value{
A named list of objects from the inference process, containing parameter samples from the inference process, the maximum likelihood parameterisation, likelihood samples, and the sampling times.
}
\description{
Runs HyperTraPS-related inference on a dataset of observations
}
