% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\docType{package}
\name{hypertrapsct}
\alias{hypertrapsct-package}
\alias{hypertrapsct}
\title{Hypercubic inference in continuous time}
\description{
Something
}
\author{
\strong{Maintainer}: Johnston Iain \email{iain.johnston@uib.no} (\href{https://orcid.org/0000-0001-8559-3519}{ORCID})

Authors:
\itemize{
  \item Aga Olav (\href{https://orcid.org/0009-0004-1610-9403}{ORCID})
  \item Brun Morten
  \item Dauda Kazeem (\href{https://orcid.org/0000-0002-4392-8592}{ORCID})
  \item Diaz-Uriarte Ramon (\href{https://orcid.org/0000-0002-6637-9039}{ORCID})
  \item Giannakis Konstantinos (\href{https://orcid.org/0000-0003-4199-6708}{ORCID})
}

}
