% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.bubbles.coarse}
\alias{plotHypercube.bubbles.coarse}
\title{Binned acquisition ordering}
\usage{
plotHypercube.bubbles.coarse(
  my.post,
  reorder = FALSE,
  transpose = FALSE,
  bins = 5
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{reorder}{Logical. Defaults to FALSE.}

\item{transpose}{Logical. Defaults to FALSE.}

\item{bins}{Numeric. Defaults to 5.}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles.coarse(fitted.cube)
}
