% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.bubbles}
\alias{plotHypercube.bubbles}
\title{Acquisition ordering}
\usage{
plotHypercube.bubbles(my.post, reorder = FALSE, transpose = FALSE)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles(fitted.cube)
}
