% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.curated.tree}
\alias{plotHypercube.curated.tree}
\title{Visualize phylogenetic tree with features}
\usage{
plotHypercube.curated.tree(tree.set)
}
\arguments{
\item{curated.tree}{combined phylogeny and feature data created with [curate.tree()]}
}
\value{
a ggplot
}
\description{
Creates a plot
}
