% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.graph}
\alias{plotHypercube.graph}
\title{Plot transition graph}
\usage{
plotHypercube.graph(
  my.post,
  thresh = 0.05,
  node.labels = TRUE,
  node.label.size = 2,
  node.labels.box = FALSE
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{thresh}{A numeric from 0 to 1. Lower flux threshold to display. Defaults to 0.05.}

\item{node.labels}{Logical. Defaults to TUE}

\item{node.label.size}{A numeric. Defaults to 2}

\item{node.label.box}{A logical. Defaults to FALSE.}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.graph(fitted.cube)
}
