% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.influencegraph}
\alias{plotHypercube.influencegraph}
\title{Visualize pairwise influences}
\usage{
plotHypercube.influencegraph(
  my.post,
  featurenames = c(""),
  use.regularised = FALSE,
  use.final = FALSE,
  thresh = 0.05,
  cv.thresh = Inf,
  label.size = 2,
  red.green = FALSE
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS.}

\item{featurenames}{A character vector of length L. Optional}

\item{use.regularised}{A logical. Defaults to FALSE}

\item{use.final}{A logical. Defaults to FALSE.}

\item{thresh}{A numeric describing the threshold. Defaults to 0.05}

\item{cv.thresh}{A numeric describing the upper bound. Defaults to Inf}

\item{label.size}{A numeric describing label size in mm. Defaults to 2}

\item{red.green}{A logical.}
}
\value{
a ggplot
}
\description{
plot pairwise influences between features in the L^2 or L^3 picture
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.influencegraph(fitted.cube)
}
