% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.influences}
\alias{plotHypercube.influences}
\title{Visualize pairwise influences}
\usage{
plotHypercube.influences(
  my.post,
  featurenames = c(""),
  use.regularised = FALSE,
  use.final = FALSE,
  reorder = FALSE,
  upper.right = FALSE,
  cv.thresh = Inf,
  red.green = FALSE
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{featurenames}{A character vector.}

\item{use.regularised}{Logical. Defaults to FALSE.}

\item{use.final}{Logical. Defaults to FALSE.}

\item{reorder}{Logical. Defaults to FALSE.}

\item{upper.right}{Logical. Defaults to FALSE.}

\item{cv.thresh}{Numeric. Upper threshold. Defaults to Inf.}

\item{red.green}{Logical. Adjust colors for red-green color-blindness. 
Defaults to FALSE.}
}
\value{
a ggplot
}
\description{
Plot pairwise influences between features in the L^2 picture.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles(fitted.cube)
}
