% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.lik.trace}
\alias{plotHypercube.lik.trace}
\title{Likelihood trace}
\usage{
plotHypercube.lik.trace(my.post)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.lik.trace(fitted.cube)
}
