% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.motifs}
\alias{plotHypercube.motifs}
\title{Motifs}
\usage{
plotHypercube.motifs(
  my.post,
  featurenames = c(""),
  label.size = 3,
  label.scheme = "full"
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{featurenames}{Optional character vector}

\item{label.size}{Numeric. Defaults to 3.}

\item{label.scheme}{String. Defaults to "full". Will remove the labels if any
other value is passed.}
}
\value{
a ggplot
}
\description{
Summarize the inferred acquisition order. Analogous to [plotHypercube.bubbles()]
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.motifs(fitted.cube)
}
