% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.motifseries}
\alias{plotHypercube.motifseries}
\title{Visualize probabilities at a given set of times}
\usage{
plotHypercube.motifseries(my.post, t.set = 0, thresh = 0.05, label.size = 3)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS.}

\item{t.set}{A numeric vector of times to visualize.}

\item{thresh}{A numeric of the lower bound threshold for probabilities 
to display. Defaults to 0.05.}

\item{label.size}{Defaults to 3.}
}
\value{
a ggplot
}
\description{
Creates a plot
plot genotype probabilities at a given set of times as a "motif" plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.motifseries(fitted.cube)
}
