% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.prediction}
\alias{plotHypercube.prediction}
\title{Visualize predicted hidden values}
\usage{
plotHypercube.prediction(prediction, max.size = 30)
}
\arguments{
\item{prediction}{Created with `predictHiddenVals`}

\item{max.size}{Defaults to 30. Size of largest points}
}
\value{
a ggplot
}
\description{
`plotHypercube.prediction` visualizes output from `predictHiddenVals`
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
prediction <- predictHiddenVals(fitted.cube, c(0,2,1))
plotHypercube.predicton(prediction)
}
