% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.sampledgraph}
\alias{plotHypercube.sampledgraph}
\title{Transition graph plot}
\usage{
plotHypercube.sampledgraph(
  my.post,
  max.samps = 1000,
  thresh = 0.05,
  node.labels = TRUE,
  node.label.size = 2
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{max.samps}{A numeric. Number of trajectories to plot. Defaults to 1000.}

\item{thresh}{A numeric from 0 to 1. Lower flux threshold to display. Defaults to 0.05.}

\item{node.labels}{Logical. Whether to display node labels. Defaults to TRUE.}

\item{node.label.size}{A numeric. Defaults to 2}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.sampledgraph(fitted.cube)
}
