% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.sampledgraph2}
\alias{plotHypercube.sampledgraph2}
\title{Transition graph plot}
\usage{
plotHypercube.sampledgraph2(
  my.post,
  max.samps = 1000,
  thresh = 0.05,
  node.labels = TRUE,
  use.arc = FALSE,
  no.times = FALSE,
  small.times = FALSE,
  times.offset = c(0.1, -0.1),
  edge.label.size = 2,
  edge.label.angle = "across",
  edge.label.colour = "#000000",
  edge.check.overlap = TRUE,
  featurenames = TRUE,
  truncate = -1,
  node.label.size = 2,
  use.timediffs = TRUE
)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{max.samps}{A numeric. Number of trajectories to plot. Defaults to 1000.}

\item{thresh}{A numeric from 0 to 1. Lower flux threshold to display. Defaults to 0.05.}

\item{node.labels}{Logical. Whether to display node labels. Defaults to TRUE.}

\item{use.arc}{Logical. Whether to use curved edges. Defaults to FALSE.}

\item{no.times}{Logical. Whether to display timing on edges. Defaults to FALSE.}

\item{small.times}{Logical. Whether to make the timing information tiny.}

\item{times.offset}{A numeric vector with relative x and y coordinates. 
Defaults to c(0.1, -0.1).}

\item{edge.label.size}{A numeric. Defaults to 2.}

\item{edge.label.angle}{A string. Unused}

\item{edge.label.colour}{A hex color code. Defaults to "#000000".}

\item{edge.check.overlap}{A logical. Passed to [ggraph::geom_edge_link()]. 
If TRUE, text that overlaps previous text in the same layer will not be plotted. 
Defaults to TRUE.}

\item{featurenames}{A character vector}

\item{truncate}{An optional integer. Limits the number of features plotted.}

\item{node.label.size}{A numeric. Defaults to 2}

\item{use.timediffs}{Logical. Defaults to TRUE.}
}
\value{
a ggplot
}
\description{
Creates a plot
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.sampledgraph2(fitted.cube, no.times = TRUE, use.timediffs = FALSE)
}
