% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PosteriorAnalysis}
\alias{PosteriorAnalysis}
\title{Extracts information from HyperTraPS-related posterior samples}
\usage{
PosteriorAnalysis(
  L,
  featurenames = NULL,
  use_regularised = 0L,
  limited_output = 0L,
  samples_per_row = 10L
)
}
\arguments{
\item{L}{List output from HyperTraPS, containing posterior samples}
}
\value{
Named list containing summary data for feature acquisition ordering ("bubbles"), time histograms, sampled accumulation routes, and timings of these sampled routes.
}
\description{
Extracts information from HyperTraPS-related posterior samples
}
