% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{predictHiddenVals}
\alias{predictHiddenVals}
\title{Impute data}
\usage{
predictHiddenVals(my.post, state, level.weight = 1)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{state}{An integer vector detailing the state to impute. Hidden values 
are denoted with 2.}
}
\value{
A list containing two data frames: state.probs and locus.probs.
}
\description{
predict unobserved values in a given observation
only works so far if the posterior structure has transition dynamics information
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
predictHiddenVals(fitted.cube, c(0,1,2))
}
