% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{predictNextStep}
\alias{predictNextStep}
\title{Predict the next likely step}
\usage{
predictNextStep(my.post, state)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS}

\item{state}{A numeric vector of 1s and 0s.}
}
\value{
A data frame of states and probabilities for the next step
}
\description{
predict the next evolutionary step from a given state
only works so far if the posterior structure has transition dynamics information
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
predictNextStep(fitted.cube, c(0,0,1))
}
