% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{readHyperinf}
\alias{readHyperinf}
\title{Read data written to file}
\usage{
readHyperinf(label, postlabel = "", fulloutput = FALSE, regularised = FALSE)
}
\arguments{
\item{label}{A string that describes the path and prefix that will be used.}

\item{postlabel}{A string that describes the path and prefix that will be 
used to save posterior analysis information to file.}

\item{fulloutput}{whether states and transitions should be read.
Defaults to FALSE.}

\item{regularised}{Whether regularisation data should be read.
Defaults to FALSE.}
}
\value{
A list of loaded data.
}
\description{
`readHyperinf` reads csv-files created by [writeHyperinf()].
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles(fitted.cube)
}
