% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{writeHyperinf}
\alias{writeHyperinf}
\title{Save a fitted hypercube}
\usage{
writeHyperinf(
  wL,
  label,
  postlabel = "",
  fulloutput = FALSE,
  regularised = FALSE
)
}
\arguments{
\item{wL}{A fitted hypercube returned from [HyperTraPS()]}

\item{label}{A string that describes the path and prefix that will be used.}

\item{postlabel}{A string that describes the path and prefix that will be 
used to save posterior analysis information to file.}

\item{fulloutput}{whether states and transitions should be written as
csv-files. Defaults to FALSE.}

\item{regularised}{Whether regularisation data should be written to file.
Defaults to FALSE.}
}
\description{
`writeHyperinf` saves output from [HyperTraPS()] to file.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
writeHyperInf(fitted.cube, "my_cube")
}
