% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.timeseries}
\alias{plotHypercube.timeseries}
\title{Time series plot}
\usage{
plotHypercube.timeseries(my.post, log.time = TRUE, featurenames = TRUE)
}
\arguments{
\item{my.post}{A fitted hypercube returned from HyperTraPS.}

\item{log.time}{Whether or not to log-transform timings. Defaults to TRUE}

\item{featurenames}{A logical. Whether or not to use feature names. Defaults to TRUE}
}
\value{
a ggplot
}
\description{
Visualize the acquisition of features in continuous time.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.timeseries(fitted.cube)
}
